requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
torch
torchvision
scikit-learn
ultralytics
streamlit